+++ src/components/MobileSidebar.tsx
@@ -1,13 +1,15 @@
-import { HamburgerMenuIcon } from "@radix-ui/react-icons";
+import { HistoryEntryType } from "@/lib/common";
 import { XIcon } from "lucide-react";
-import { MouseEvent, ReactNode, useState } from "react";
+import { MouseEvent, ReactNode } from "react";
 import { Button } from "./ui/button";
 import Sidebar from "./ui/sidebar";
-import { HistoryEntryType } from "@/lib/common";
 
 export default function MobileSidebar({
 	isHistoryOpen,
+	setHistoryOpen,
+	sidebarIsOpen,
 	setSidebarisOpen,
+	clearResultHistory,
 	newRequestHandler,
 	openSplashScreen,
 	children,
@@ -15,18 +17,19 @@ export default function MobileSidebar({
 	resultHistory: HistoryEntryType[];
 	restoreResultHistoryItem: (id: string) => void;
 	isHistoryOpen: boolean;
+	setHistoryOpen: (_: boolean) => void;
+	sidebarIsOpen: boolean;
 	setSidebarisOpen: (isOpen: boolean) => void;
+	clearResultHistory: () => void;
 	newRequestHandler: (event: MouseEvent<HTMLButtonElement>) => void;
-	openSplashScreen:() => void;
+	openSplashScreen: () => void;
 	children: ReactNode;
 }) {
-	const [isMobileSidebarVisible, setIsMobileSidebarVisible] = useState(false);
-
 	return (
 		<>
 			<div
 				className={`absolute lg:hidden top-0 left-0 w-screen h-screen bg-grey-300 bg-slate-400 bg-opacity-40 z-40 ${
-					isMobileSidebarVisible ? "visible" : "invisible"
+					sidebarIsOpen ? "visible" : "invisible"
 				}`}
 				style={{
 					maxHeight: "-webkit-fill-available",
@@ -36,15 +39,18 @@ export default function MobileSidebar({
 				<div className="flex w-full h-full">
 					<div
 						className={`flex items-start w-80 h-full ease-in-out duration-300 ${
-							isMobileSidebarVisible ? "translate-x-0" : "-translate-x-80"
+							sidebarIsOpen ? "translate-x-0" : "-translate-x-80"
 						}`}
 					>
 						<aside className="w-80 h-screen border-r overflow-auto bg-white px-4 shadow-lg pb-4">
 							<Sidebar
-								sidebarIsOpen={isHistoryOpen}
+								sidebarIsOpen={sidebarIsOpen}
 								onNewRequest={newRequestHandler}
-								onSidebarOpenChange={setSidebarisOpen}
+								setSidebarIsOpen={setSidebarisOpen}
 								openSplashScreen={openSplashScreen}
+								clearResultHistory={clearResultHistory}
+								setHistoryIsOpen={setHistoryOpen}
+								historyIsOpen={isHistoryOpen}
 							>
 								{children}
 							</Sidebar>
@@ -53,12 +59,12 @@ export default function MobileSidebar({
 							className={[
 								"px-2 sm:px-3 py-3 shadow-md flex gap-2 absolute top-2 left-[calc(100%+0.5rem)]",
 								"transition-opacity",
-								!isMobileSidebarVisible && "opacity-0 pointer-events-none",
-								isMobileSidebarVisible && "opacity-100",
+								!sidebarIsOpen && "opacity-0 pointer-events-none",
+								sidebarIsOpen && "opacity-100",
 							]
 								.filter(Boolean)
 								.join(" ")}
-							onClick={() => setIsMobileSidebarVisible(false)}
+							onClick={() => setSidebarisOpen(false)}
 						>
 							<XIcon />
 							<span className="hidden sm:inline">Schließen</span>
@@ -66,18 +72,9 @@ export default function MobileSidebar({
 					</div>
 					<div
 						className="flex w-1 grow"
-						onClick={() => setIsMobileSidebarVisible(false)}
+						onClick={() => setSidebarisOpen(false)}
 					></div>
 				</div>
-			</div>
-			<div className="absolute top-0 lg:hidden z-60 p-3">
-				<Button
-					onClick={() => setIsMobileSidebarVisible(!isMobileSidebarVisible)}
-					className="p3"
-					size="icon"
-				>
-					<HamburgerMenuIcon width={20} height={20} />
-				</Button>
 			</div>
 		</>
 	);