+++ src/components/SearchResultSection.tsx
@@ -1,9 +1,9 @@
-import * as Tooltip from "@radix-ui/react-tooltip";
 import type { ResponseDocumentMatch } from "@/lib/common";
-import React, { ReactNode, useState } from "react";
-import { Link } from "./Link";
 import { cn, getCleanedMetadata } from "@/lib/utils";
+import { useState } from "react";
+import { Link } from "./Link";
 import { AcrobatIcon } from "./ui/acrobat-icon";
+import { GlobeIcon } from "@radix-ui/react-icons";
 
 interface SearchResultProps {
 	documentMatch: ResponseDocumentMatch | undefined;
@@ -21,14 +21,13 @@ function ExpandableTableCell({ content }: { content: string }) {
 	return (
 		<p className="text-sm sm:text-base mb-4">
 			{displayedContent}
-			<br />
 			<button
 				className={cn(
 					"underline text-blue-700 hover:text-blue-900",
 					"bg-none focus-visible:outline-none",
 					"focus-visible:ring-2 focus-visible:ring-blue-700",
 					"focus-visible:ring-offset-4 focus-visible:ring-offset-white",
-					"focus-visible:rounded-sm",
+					"focus-visible:rounded-sm ml-2",
 				)}
 				onClick={handleClick}
 			>
@@ -38,10 +37,11 @@ function ExpandableTableCell({ content }: { content: string }) {
 	);
 }
 
-function TagsList(props: { tags: string[] }) {
-	const [isExpanded, setIsExpanded] = useState<boolean>(false);
-	const maxTags = 3;
-	const tags = isExpanded ? props.tags : props.tags.slice(0, maxTags);
+interface TagsListProps {
+	tags: string[];
+}
+
+function TagsList({ tags }: TagsListProps) {
 	return (
 		<div className="flex gap-x-2 gap-y-1 flex-wrap">
 			{tags.map((tag) => (
@@ -52,23 +52,6 @@ function TagsList(props: { tags: string[] }) {
 					{tag}
 				</span>
 			))}
-			{props.tags.length > maxTags && (
-				<button
-					className={cn(
-						"inline-block rounded-full transition-colors text-xs",
-						isExpanded
-							? "text-blue-700 underline"
-							: "bg-slate-100 text-slate-600 px-2 py-0.5 hover:bg-blue-900 hover:text-white",
-						"focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-blue-700",
-						"focus-visible:ring-offset-4 focus-visible:ring-offset-white",
-					)}
-					onClick={() => setIsExpanded(!isExpanded)}
-				>
-					{!isExpanded
-						? `+${props.tags.length - maxTags}`
-						: `Schlagworte ausblenden`}
-				</button>
-			)}
 		</div>
 	);
 }
@@ -77,7 +60,7 @@ type SimilarityDisplayProps = {
 	similarity: number;
 };
 
-function SimilarityDisplay(props: SimilarityDisplayProps): ReactNode {
+function SimilarityDisplay(props: SimilarityDisplayProps): JSX.Element {
 	const similarityRouded = Math.floor(props.similarity * 1000) / 10;
 	return <span className="text-xs">{similarityRouded}% Relevanz</span>;
 }
@@ -85,28 +68,47 @@ function SimilarityDisplay(props: SimilarityDisplayProps): ReactNode {
 export default function SearchResultSection({
 	documentMatch,
 }: SearchResultProps) {
-	const { title, pdfUrl, pdfName, pages, similarity, type, tags } =
+	const { title, pdfUrl, documentName, pages, similarity, type, tags } =
 		getCleanedMetadata(documentMatch);
 
+	const isWebSource = type === "Webseite";
+
+	// Hacky fix: "Hauptausschussprotokoll" documents in the database should be named "Hauptausschussvorgang"
+	// Ideally, this should be fixed in the database
+	const documenType =
+		type === "Hauptausschussprotokoll" ? "Hauptausschussvorgang" : type;
+
+	const processedAt = new Date(
+		documentMatch?.processed_document.processing_finished_at!,
+	);
 	return (
-		<div className="border-x border-t last-of-type:border-b first-of-type:rounded-t last-of-type:rounded-b p-4">
-			<div className="flex gap-2 justify-between items-center relative">
+		<div className="bg-white p-4 rounded-lg shadow-md">
+			<div className="flex gap-2 justify-between items-center">
 				<span className="flex gap-x-2 flex-wrap items-center text-sm sm:text-base">
-					<span>{type}</span>
+					<span>{documenType}</span>
 					<span className="text-slate-400 hidden sm:inline" aria-hidden="true">
 						∙
 					</span>
 					{pdfUrl && (
 						<Link
 							href={pdfUrl}
-							title={`PDF "${pdfName}" öffnen`}
+							title={`Dokument "${documentName}" öffnen`}
 							className="no-underline flex gap-1 items-center flex-wrap"
 						>
-							<AcrobatIcon />
-							<span>{pdfName}</span>{" "}
-							{pages && pages.length > 0 && (
+							{pdfUrl.endsWith(".pdf") ? (
+								<AcrobatIcon />
+							) : (
+								<GlobeIcon></GlobeIcon>
+							)}
+							<span>{documentName}</span>{" "}
+							{!isWebSource && pages && pages.length > 0 && (
 								<span className="text-slate-600 ml-1">
 									(S. {pages?.join(", ")})
+								</span>
+							)}
+							{isWebSource && (
+								<span className="text-slate-600 ml-1">
+									(Abgerufen am {processedAt.toLocaleDateString()})
 								</span>
 							)}
 						</Link>
@@ -115,9 +117,7 @@ export default function SearchResultSection({
 				<SimilarityDisplay similarity={similarity} />
 			</div>
 
-			<h6 className="font-bold text-lg mt-2 sm:mt-1 mb-3 leading-snug">
-				{title}
-			</h6>
+			<h6 className="text-lg mt-2 sm:mt-1 mb-3 leading-snug">{title}</h6>
 
 			<ExpandableTableCell
 				content={