+++ src/components/ui/answer.tsx
@@ -1,12 +1,10 @@
 import { DocumentSearchResponse } from "@/lib/common";
-import React, { ReactNode, useEffect, useState } from "react";
-import SearchResultSection from "../SearchResultSection";
-import AnswerLoadingSkeleton from "./textLoadingSkeleton";
-import DocumentLoadingSkeleton from "./documentLoadingSkeleton";
-import ReactMarkdown from "react-markdown";
-import { cn, getCleanedMetadata } from "@/lib/utils";
 import { getDocumentsCount } from "@/lib/get-documents-count";
 import { texts } from "@/lib/texts";
+import { cn } from "@/lib/utils";
+import { useEffect, useState } from "react";
+import ReactMarkdown from "react-markdown";
+import AnswerLoadingSkeleton from "./textLoadingSkeleton";
 
 type AnswerProps = {
 	generatedAnswer: string | null;
@@ -17,7 +15,7 @@ type AnswerProps = {
 
 const formatter = new Intl.NumberFormat("de-DE");
 
-function Answer(props: AnswerProps): ReactNode {
+function Answer(props: AnswerProps): JSX.Element {
 	const [documentsCount, setDocumentsCount] = useState("");
 	const { generatedAnswer, searchResult, searchIsLoading, answerIsLoading } =
 		props;
@@ -32,33 +30,46 @@ function Answer(props: AnswerProps): ReactNode {
 	if (!searchIsLoading && searchResult && matches.length === 0) {
 		return (
 			<>
-				<h4 className="text-lg font-bold mb-2">{texts.noResultsTitle}</h4>
+				<h4 className="mb-2 leading-6 bg-white p-4 rounded-lg shadow-md">
+					{texts.noResultsTitle}
+				</h4>
 			</>
 		);
 	}
 
 	return (
 		<>
-			<div className="mb-4">
-				<h5 className="text-lg font-bold mb-2">
+			<div>
+				<h5 className="text-lg mb-2 ml-4">
 					{answerIsLoading && texts.answerIsLoading}
 					{!answerIsLoading && generatedAnswer && texts.answerTitle}
 				</h5>
-				{answerIsLoading && !generatedAnswer && <AnswerLoadingSkeleton />}
+				{answerIsLoading && !generatedAnswer && (
+					<div
+						data-testid="answer-loading-skeleton"
+						className="rounded-lg shadow-md p-4 bg-white"
+					>
+						<AnswerLoadingSkeleton />
+					</div>
+				)}
 				{generatedAnswer && (
 					<>
-						<ReactMarkdown className="prose leading-6">
-							{generatedAnswer}
-						</ReactMarkdown>
+						<div data-testid="generated-answer">
+							<ReactMarkdown className="markdown-container leading-6 bg-white p-4 rounded-lg shadow-md space-y-4">
+								{generatedAnswer}
+							</ReactMarkdown>
+						</div>
+
 						<div
 							tabIndex={0}
 							className={cn(
-								`prose leading-6 focus:outline-none group`,
-								`mt-6 border-t border-slate-200 mb-8 text-sm`,
+								`leading-6 pt-4 pl-4 pr-4 pb-0 rounded-lg text-sm w-full`,
 							)}
+							data-testid="generated-answer-disclaimer"
 						>
 							<ReactMarkdown
 								className={cn(
+									`markdown-container`,
 									`text-slate-500 transition-colors group-hover:text-slate-900`,
 									`group-focus-visible:text-slate-900 prose-strong:text-inherit`,
 									`group-focus-visible:ring-2 group-focus-visible:rounded-sm`,
@@ -72,31 +83,7 @@ ${texts.answerDisclaimer}
 						</div>
 					</>
 				)}
-				<h5 className="font-bold mt-4">
-					{searchIsLoading &&
-						`${documentsCount} ${texts.documentsAreLoading}.`.trim()}
-					{!searchIsLoading && searchResult && texts.documentsTitle}
-				</h5>
 			</div>
-			{searchIsLoading && <DocumentLoadingSkeleton />}
-			{!searchIsLoading && matches.length > 0 && (
-				<div className="w-[calc(100%+2rem)] -ml-4">
-					{matches
-						.sort((l, r) => {
-							const lm = getCleanedMetadata(l);
-							const rm = getCleanedMetadata(r);
-							return lm.similarity < rm.similarity ? 1 : -1;
-						})
-						.map((documentMatch) => {
-							return (
-								<SearchResultSection
-									key={documentMatch.registered_document.id}
-									documentMatch={documentMatch}
-								/>
-							);
-						})}
-				</div>
-			)}
 		</>
 	);
 }