+++ src/components/ui/promptForm.tsx
@@ -1,40 +1,71 @@
-import React, { FormEventHandler, ReactNode } from "react";
-import { Input } from "./input";
-import { Button } from "./button";
-import Envelope from "./envelopeIcon";
-import { cn } from "@/lib/utils";
+import { MagnifyingGlassIcon } from "@radix-ui/react-icons";
+import React, { useEffect, useRef, useState } from "react";
 
 type PromptFormProps = {
-	onSubmit: (query?: string) => void;
-	query?: string;
+	onSubmit: (query: string | null) => void;
+	query: string | null;
 	isLoading: boolean;
 	onChange: (event: React.ChangeEvent<HTMLTextAreaElement>) => void;
 };
 
-function PromptForm(props: PromptFormProps): ReactNode {
+function PromptForm(props: PromptFormProps): JSX.Element {
 	const { onSubmit, query, isLoading, onChange } = props;
+	const textAreaRef = useRef<HTMLTextAreaElement>(null);
+	const [uiInitializing, setUiInitializing] = useState(true);
+
+	useEffect(() => {
+		handleInputChange();
+		setUiInitializing(false);
+		window.addEventListener(
+			"resize",
+			() => {
+				handleInputChange();
+			},
+			false,
+		);
+	}, []);
+
+	useEffect(() => {
+		handleInputChange();
+	}, [query]);
+
+	// See https://stackoverflow.com/questions/2803880/is-there-a-way-to-get-a-textarea-to-stretch-to-fit-its-content-without-using-php for reference
+	const handleInputChange = () => {
+		const textArea = textAreaRef.current;
+		if (textArea) {
+			if (!textArea.value || textArea.value === "") {
+				textArea.value =
+					"Stellen Sie hier Ihre Frage oder wählen Sie eines der Beispiele aus";
+				textArea.style.height = ""; // Reset the height to recalculate the scroll height
+				textArea.style.height = `${textArea.scrollHeight}px`; // Set the height based on scrollHeight
+				textArea.value = "";
+			} else {
+				textArea.style.height = ""; // Reset the height to recalculate the scroll height
+				textArea.style.height = `${textArea.scrollHeight}px`; // Set the height based on scrollHeight
+			}
+		}
+	};
+
 	return (
-		<div
-			className={cn(
-				"w-full sticky bottom-0",
-				"p-2 lg:px-0 bg-slate-50 border-t border-slate-200 shadow-lg",
-			)}
-		>
+		<div className={`${uiInitializing ? "opacity-0" : ""}`}>
+			<div className="w-full h-0 sm:h-12 mt-2 sm:mt-0 sm:h-6 bg-[#F8FAFC]"></div>
 			<form
 				onSubmit={(evt) => {
 					evt.preventDefault();
 					onSubmit(query);
 				}}
-				className="relative w-full max-w-xl mx-auto flex"
+				className={`relative w-full max-w-3xl mx-auto h-auto`}
 				name="promptForm"
 			>
-				<Input
+				<textarea
+					data-testid="prompt-textarea"
+					ref={textAreaRef}
 					value={query || ""}
+					rows={1}
 					name="query"
 					id="query"
-					className="pl-4 py-4 pr-12 resize-none"
-					placeholder="Stellen Sie hier Ihre Frage"
-					onChange={onChange}
+					className="w-full py-4 pl-4 pr-[45px] rounded-md shadow-md shadow-blue-200 resize-none h-auto"
+					placeholder="Stellen Sie hier Ihre Frage oder wählen Sie eines der Beispiele aus"
 					disabled={isLoading}
 					onKeyDown={(event) => {
 						if (event.key === "Enter" && !event.shiftKey) {
@@ -42,15 +73,27 @@ function PromptForm(props: PromptFormProps): ReactNode {
 							onSubmit(query);
 						}
 					}}
+					onChange={(e) => {
+						handleInputChange();
+						onChange(e);
+					}}
 				/>
-				<Button
-					className="absolute right-2 bottom-2 bg-blue-700 hover:bg-blue-900 text-white font-bold"
+				<button
+					className="absolute right-4 top-4 text-white"
 					type="submit"
-					size="icon"
 					disabled={isLoading}
 				>
-					{isLoading ? <Spinner /> : <Envelope />}
-				</Button>
+					{isLoading ? (
+						<Spinner />
+					) : (
+						<div>
+							<MagnifyingGlassIcon
+								className="w-6 h-6"
+								color="gray"
+							></MagnifyingGlassIcon>
+						</div>
+					)}
+				</button>
 			</form>
 		</div>
 	);
@@ -61,7 +104,7 @@ function Spinner() {
 		<svg
 			aria-hidden="true"
 			role="status"
-			className="inline w-4 h-4 text-gray-200 animate-spin dark:text-gray-600"
+			className="inline w-5 h-5 text-gray-200 animate-spin dark:text-gray-600"
 			viewBox="0 0 100 101"
 			fill="none"
 			xmlns="http://www.w3.org/2000/svg"