+++ src/components/ui/promtContent.tsx
@@ -4,6 +4,8 @@ import Answer from "./answer";
 import ExamplePrompts from "./examplePrompts";
 import exampleQuestions from "@/fixtures/example-questions";
 import { selectRandomItems } from "@/lib/utils";
+import TypeformLink from "./TypeformLink";
+import { texts } from "@/lib/texts";
 
 type PromptContentProps = {
 	title?: string | null;
@@ -34,27 +36,12 @@ function PromptContent(props: PromptContentProps) {
 	>([]);
 
 	useEffect(() => {
-		setExamplesQuestionsToShow(selectRandomItems(exampleQuestions, 3));
+		setExamplesQuestionsToShow(selectRandomItems(exampleQuestions, 5));
 	}, []);
 
 	return (
-		<div className="space-y-2 pt-8 lg:pt-0">
-			{!title && (
-				<div className="max-w-xl mx-auto">
-					<h3 className="leading-snug text-xl py-3">
-						Schreiben Sie Ihre Frage in das Suchfeld oder klicken Sie auf auf
-						eines der Beispiele, um den KI-Assistenten zu starten.
-					</h3>
-				</div>
-			)}
-			<div className="max-w-xl mx-auto">
-				{title && (
-					<>
-						<p className="leading-7 text-lg font-light whitespace-pre-wrap mb-6">
-							{title}
-						</p>
-					</>
-				)}
+		<div className="">
+			<div className="max-w-3xl mx-auto">
 				<Answer
 					generatedAnswer={generatedAnswer}
 					answerIsLoading={answerIsLoading}
@@ -62,11 +49,17 @@ function PromptContent(props: PromptContentProps) {
 					searchIsLoading={searchIsLoading}
 				/>
 			</div>
-			{showExamplePrompts && (
-				<ExamplePrompts
-					examplePrompts={exampleQuestionsToShow}
-					onClick={(text) => onsubmit(text)}
-				/>
+			{showExamplePrompts && exampleQuestionsToShow.length > 0 && (
+				<>
+					<ExamplePrompts
+						examplePrompts={exampleQuestionsToShow}
+						onClick={(text) => onsubmit(text)}
+					/>
+					<TypeformLink
+						question={texts.feedback.question}
+						linkText={texts.feedback.long}
+					></TypeformLink>
+				</>
 			)}
 		</div>
 	);