+++ src/components/ui/sidebar.tsx
@@ -1,79 +1,78 @@
-import { cn } from "@/lib/utils";
-import { MouseEventHandler, ReactNode } from "react";
-import { Button } from "./button";
-import { ChevronDownIcon, ChevronLeftIcon } from "@radix-ui/react-icons";
 import { Collapsible, CollapsibleContent } from "@radix-ui/react-collapsible";
+import {
+	ChevronDownIcon,
+	ChevronRightIcon,
+	InfoCircledIcon,
+	PlusCircledIcon,
+} from "@radix-ui/react-icons";
+import { MouseEventHandler, ReactNode } from "react";
 
 type SidebarProps = {
-	onNewRequest: MouseEventHandler<HTMLButtonElement>;
-	onSidebarOpenChange: (open: boolean) => void;
-	openSplashScreen: () => void;
 	sidebarIsOpen: boolean;
+	setSidebarIsOpen: (open: boolean) => void;
+	historyIsOpen: boolean;
+	setHistoryIsOpen: (open: boolean) => void;
+	clearResultHistory: () => void;
+	onNewRequest: MouseEventHandler<HTMLButtonElement>;
+	openSplashScreen: () => void;
 	children: ReactNode;
 };
 
 function Sidebar(props: SidebarProps): JSX.Element {
 	return (
 		<>
-			<header className="flex items-center justify-between px-1 py-2">
-				<a
-					href="/"
-					title="Startseite"
-					className={cn(
-						"py-3 px-4 flex justify-between items-center text-lg font-bold rounded",
-						"hover:text-slate-600",
-						"focus-visible:ring-2 focus-visible:ring-blue-700",
-						"focus-visible:outline-none",
-					)}
-				>
-					Parla
-				</a>
+			<div className="sticky top-0 left-0 bg-inherit">
+				<header className="flex items-center justify-between px-4 pb-7 pt-3">
+					<a href="/">
+						<img
+							alt="Parla Logo"
+							src="/images/parla-logo-v1.svg"
+							className="w-[50%] py-3"
+						/>
+					</a>
+					<button onClick={props.openSplashScreen}>
+						<InfoCircledIcon className="w-6 h-6 text-slate-400"></InfoCircledIcon>
+					</button>
+				</header>
+
 				<button
-					className={cn(
-						"px-2 hover:bg-blue-900",
-						"text-sm text-slate-400 hover:text-white",
-						"border-2 rounded-full border-slate-400 hover:border-blue-900",
-						"focus-visible:ring-2 focus-visible:ring-blue-700 focus-visible:ring-offset-2",
-						"focus-visible:outline-none",
-					)}
-					onClick={props.openSplashScreen}
-				>
-					i
-				</button>
-			</header>
-			<div className="px-1 py-2 border-y border-slate-200">
-				<Button
+					className="w-full flex items-center justify-between px-4 py-2 rounded-md bg-blue-700 text-white font-bold hover:bg-blue-900 hover:cursor-pointer"
 					onClick={props.onNewRequest}
-					className="w-full text-white bg-blue-700 hover:bg-blue-900 font-bold py-2 px-4 flex justify-between"
 				>
-					<span>Neue Frage</span>
-					<span>+</span>
-				</Button>
-			</div>
-			<div className="px-1 py-2">
-				<button
-					className={cn(
-						"flex bg-inherit justify-between w-full items-center hover:bg-none px-4 py-3",
-						"focus-visible:ring-2 focus-visible:ring-blue-700",
-						"focus-visible:outline-none focus-visible:rounded-sm my-2",
-					)}
-					onClick={() => props.onSidebarOpenChange(!props.sidebarIsOpen)}
-				>
-					<strong className="block font-bold">Vorherige Fragen</strong>
-					{props.sidebarIsOpen ? (
-						<ChevronDownIcon className="text-slate-400" />
-					) : (
-						<ChevronLeftIcon className="text-slate-400" />
-					)}
+					<div>Neue Frage</div>
+					<div>
+						<PlusCircledIcon className="w-6 h-6 text-white"></PlusCircledIcon>
+					</div>
 				</button>
 
-				<Collapsible
-					open={props.sidebarIsOpen}
-					onOpenChange={props.onSidebarOpenChange}
+				<button
+					className="w-full flex items-center justify-between px-4 py-6 pb-1"
+					onClick={() => props.setHistoryIsOpen(!props.historyIsOpen)}
 				>
-					<CollapsibleContent>{props.children}</CollapsibleContent>
-				</Collapsible>
+					<div>Vorherige Fragen</div>
+					<div>
+						{props.historyIsOpen ? (
+							<ChevronDownIcon className="w-6 h-6 text-slate-400" />
+						) : (
+							<ChevronRightIcon className="w-6 h-6 text-slate-400" />
+						)}
+					</div>
+				</button>
+
+				<button
+					className="text-blue-700 flex items-center px-4 py-2 text-sm hover:bg-blue-900 hover:text-white rounded-md hover:cursor-pointer mb-4"
+					onClick={() => props.clearResultHistory()}
+				>
+					<div>Fragenverlauf löschen</div>
+				</button>
 			</div>
+
+			<Collapsible
+				open={props.historyIsOpen}
+				onOpenChange={props.setHistoryIsOpen}
+			>
+				<CollapsibleContent>{props.children}</CollapsibleContent>
+			</Collapsible>
 		</>
 	);
 }